https://github.com/ShadowKnightMK4/OdinSearch/milestone/3
Should be noted, while this implies Windows, the class library itself has plans to be retargeted to be general .NET.

Note the current feature being worked up is the dup foldlder thing.  I've also added a few routines to add SearchTargets and SearchAnchors to an OdinSearch class in one call with an Ienumerable rather than ONE AT A TIME, finally.