Work on DeepDirPrune. RegEx being created before the WorkThreads spawn is done now.
--
Changes:
WorkerThreadInit() - contains the old init code that originally ran in the WorkerThreadProc(). Note that it is still programmed to call this in the event that it receives an empty list as a fallback but normal execution shouldn't have that happen.

Internal class WorkerThreadArgs now has a new entry, List which will receive the initialized RegEx stuff to compare strings. WorkerThread does assume once itself is running, that list can't change.

Please keep in mind if you depend on routines/classes marked as internal. The various classes in OdinSearch marked internal are subject to change from release to release and update